# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BhashaShield 2.0 contextual voice-fraud protection API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DemoScenario {
  id: string;
  label: string;
  description: string;
  expectedLevel: string;
  caller: string;
  transcript: string;
  voiceClass: string;
  accent?: string;
}

export interface AnalysisInput {
  scenarioId: string;
  userId: string;
  callerNumber?: string;
  channel?: string;
  transcript?: string;
  voiceClass?: string;
  offline?: boolean;
}

export interface VoiceAnalysis {
  classification: string;
  probability: number;
  confidence: number;
  label: string;
}

export interface IntentAnalysis {
  type: string;
  urgency: string;
  requestedAction: string;
  socialEngineering: boolean;
}

export interface ContextAnalysis {
  schemeMatch: boolean;
  bankingMatch: boolean;
  contextualMatch: boolean;
  matchedContext: string;
  channelVerified: boolean;
}

export interface BehaviourAnalysis {
  anomalyScore: number;
  signals: string[];
}

export interface Analysis {
  id: string;
  score: number;
  level: string;
  confidence: number;
  caller: string;
  callerNumber?: string;
  channel?: string;
  language: string;
  voice: VoiceAnalysis;
  intent: IntentAnalysis;
  context: ContextAnalysis;
  behaviour: BehaviourAnalysis;
  evidence: string[];
  explanation: string;
  recommendedAction: string;
  analysisMode: string;
  transcript?: string;
  analyzedAt: string;
}

export interface AnalysisSummary {
  id: string;
  caller: string;
  score: number;
  level: string;
  time: string;
  action: string;
}

export interface Dashboard {
  totalAnalyses: number;
  highRiskCount: number;
  protectedContextCount: number;
  averageScore: number;
  recentAnalyses: AnalysisSummary[];
  systemStatus: string;
}

export interface TrustedPerson {
  id: string;
  name: string;
  relationship: string;
  phone: string;
  verified: boolean;
  channel: string;
}

export interface TrustedPersonInput {
  name: string;
  relationship: string;
  phone: string;
  channel: string;
}

export interface DirectoryEntry {
  id: string;
  institution: string;
  number: string;
  category: string;
  purpose: string;
  source: string;
  priority: string;
  lastVerified: string;
}

export interface VerificationInput {
  institution: string;
  number?: string;
}

export interface VerificationResult {
  verified: boolean;
  institution?: string;
  officialNumber?: string;
  message: string;
}

export interface FeedbackInput {
  analysisId: string;
  outcome: string;
  note?: string;
}

export interface Feedback {
  saved: boolean;
  message: string;
}