# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BhashaShield 2.0 contextual voice-fraud protection API
 * OpenAPI spec version: 0.1.0
 */
import type { BehaviourAnalysis } from './behaviourAnalysis';
import type { ContextAnalysis } from './contextAnalysis';
import type { IntentAnalysis } from './intentAnalysis';
import type { VoiceAnalysis } from './voiceAnalysis';

export interface Analysis {
  id: string;
  score: number;
  level: string;
  confidence: number;
  caller: string;
  callerNumber?: string;
  channel?: string;
  language: string;
  voice: VoiceAnalysis;
  intent: IntentAnalysis;
  context: ContextAnalysis;
  behaviour: BehaviourAnalysis;
  evidence: string[];
  explanation: string;
  recommendedAction: string;
  analysisMode: string;
  transcript?: string;
  analyzedAt: string;
}