# lib/api-zod/src/generated/types/analysisSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BhashaShield 2.0 contextual voice-fraud protection API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisSummary {
  id: string;
  caller: string;
  score: number;
  level: string;
  time: string;
  action: string;
}