# lib/api-zod/src/generated/types/behaviourAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BhashaShield 2.0 contextual voice-fraud protection API
 * OpenAPI spec version: 0.1.0
 */

export interface BehaviourAnalysis {
  anomalyScore: number;
  signals: string[];
}