# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BhashaShield 2.0 contextual voice-fraud protection API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisSummary } from './analysisSummary';

export interface Dashboard {
  totalAnalyses: number;
  highRiskCount: number;
  protectedContextCount: number;
  averageScore: number;
  recentAnalyses: AnalysisSummary[];
  systemStatus: string;
}