# lib/api-zod/src/generated/types/demoScenario.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BhashaShield 2.0 contextual voice-fraud protection API
 * OpenAPI spec version: 0.1.0
 */

export interface DemoScenario {
  id: string;
  label: string;
  description: string;
  expectedLevel: string;
  caller: string;
  transcript: string;
  voiceClass: string;
  accent?: string;
}