# lib/api-zod/src/generated/types/directoryEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BhashaShield 2.0 contextual voice-fraud protection API
 * OpenAPI spec version: 0.1.0
 */

export interface DirectoryEntry {
  id: string;
  institution: string;
  number: string;
  category: string;
  purpose: string;
  source: string;
  priority: string;
  lastVerified: string;
}