# lib/api-zod/src/generated/types/feedbackInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BhashaShield 2.0 contextual voice-fraud protection API
 * OpenAPI spec version: 0.1.0
 */

export interface FeedbackInput {
  analysisId: string;
  outcome: string;
  note?: string;
}