# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BhashaShield 2.0 contextual voice-fraud protection API
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}