# lib/api-zod/src/generated/types/trustedPersonInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BhashaShield 2.0 contextual voice-fraud protection API
 * OpenAPI spec version: 0.1.0
 */

export interface TrustedPersonInput {
  name: string;
  relationship: string;
  phone: string;
  channel: string;
}